requirements.txt
asgiref==3.5.2
Django==4.1.2
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
pytz==2022.4
sqlparse==0.4.3

